+++ packages/rschedule/src/lib/operators/index.ts
@@ -2,6 +2,7 @@ export * from './interface';
 
 export * from './subtract.operator';
 export * from './intersection.operator';
+export * from './merge-duration.operator';
 export * from './add.operator';
 export * from './unique.operator';
 export * from './occurrence-stream';